requirements.txt
# Core dependencies for LLM QA system
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-text-splitters==0.3.8

# OpenAI API
openai==1.93.0

# Vector database
faiss-cpu==1.11.0

# Data processing
pandas==2.3.0
numpy==2.3.1

# Environment management
python-dotenv==1.1.1

# Progress bars
tqdm==4.67.1

# Type hints support
typing_extensions==4.14.0

# Additional dependencies for potential future features
pydantic==2.11.7
requests==2.32.4